Finish separating valhalla_build_transit into fetch and convert steps (#1537)
* Move transit tile PBF->GPH to valhalla_convert_transit

Move common to transitpbf.h

Provide valhalla_build_transit shell script for backwards compat

* Remove unused

* Remove

* Incorporate fix for mmap tile creation

* format.sh

* Remove
