+++ CMakeLists.txt
@@ -215,7 +215,7 @@ set(valhalla_programs valhalla_run_map_match valhalla_benchmark_loki valhalla_be
 ## Valhalla data tools
 set(valhalla_data_tools valhalla_build_statistics valhalla_ways_to_edges valhalla_validate_transit
   valhalla_benchmark_admins	valhalla_build_connectivity	valhalla_build_tiles
-  valhalla_build_admins valhalla_build_transit valhalla_fetch_transit valhalla_query_transit
+  valhalla_build_admins valhalla_convert_transit valhalla_fetch_transit valhalla_query_transit
   valhalla_build_speeds  valhalla_associate_segments	valhalla_add_predicted_traffic)
 
 ## Valhalla services
@@ -259,6 +259,7 @@ install(
   FILES
     scripts/valhalla_build_config
     scripts/valhalla_build_elevation
+    scripts/valhalla_build_transit
     scripts/valhalla_build_timezones
   DESTINATION "${CMAKE_INSTALL_BINDIR}"
   PERMISSIONS
