+++ src/mjolnir/valhalla_fetch_transit.cc
@@ -11,15 +11,12 @@
 #include <thread>
 #include <unordered_set>
 
-#include "baldr/rapidjson_utils.h"
 #include <boost/algorithm/string.hpp>
 #include <boost/filesystem/operations.hpp>
 #include <boost/format.hpp>
 #include <boost/property_tree/ptree.hpp>
 #include <boost/tokenizer.hpp>
 #include <curl/curl.h>
-#include <google/protobuf/io/coded_stream.h>
-#include <google/protobuf/io/zero_copy_stream_impl_lite.h>
 
 #include "baldr/graphconstants.h"
 #include "baldr/graphid.h"
@@ -29,10 +26,12 @@
 #include "midgard/encoded.h"
 #include "midgard/logging.h"
 #include "midgard/sequence.h"
+
 #include "mjolnir/admin.h"
 #include "mjolnir/servicedays.h"
+#include "mjolnir/transitpbf.h"
 
-#include <valhalla/proto/transit_fetch.pb.h>
+#include <valhalla/proto/transit.pb.h>
 
 using namespace boost::property_tree;
 using namespace valhalla::midgard;
@@ -237,13 +236,14 @@ std::priority_queue<weighted_tile_t> which_tiles(const ptree& pt, const std::str
         std::min(bbox.maxy(), PointLL(bbox.minx(), bbox.maxy()).MidPoint(bbox.maxpt()).second);
     bbox = AABB2<PointLL>(bbox.minx(), min_y, bbox.maxx(), max_y);
     // stop count
-    auto request = url((boost::format("/api/v1/stops?total=true&per_page=0&bbox=%1%,%2%,%3%,%4%") %
-                        bbox.minx() % bbox.miny() % bbox.maxx() % bbox.maxy())
-                           .str(),
-                       pt);
+    auto request =
+        url((boost::format("/api/v1/stop_stations?total=true&per_page=0&bbox=%1%,%2%,%3%,%4%") %
+             bbox.minx() % bbox.miny() % bbox.maxx() % bbox.maxy())
+                .str(),
+            pt);
     request += import_level;
 
-    auto stops_total = curler(request, "meta.total").get<size_t>("meta.total");
+    auto stations_total = curler(request, "meta.total").get<size_t>("meta.total");
     /*
     //route count
     request = url((boost::format("/api/v1/routes?total=true&per_page=0&bbox=%1%,%2%,%3%,%4%")
@@ -257,14 +257,14 @@ std::priority_queue<weighted_tile_t> which_tiles(const ptree& pt, const std::str
     "meta.total").get<size_t>("meta.total");
     */
     // we have anything we want it
-    if (stops_total > 0 /* || routes_total > 0|| pairs_total > 0*/) {
+    if (stations_total > 0 /* || routes_total > 0|| pairs_total > 0*/) {
       prioritized.push(
           weighted_tile_t{tile,
-                          stops_total +
+                          stations_total +
                               10 /* + routes_total * 1000 + pairs_total*/}); // TODO: factor in stop
                                                                              // pairs as well
-      LOG_INFO(GraphTile::FileSuffix(tile) + " should have " + std::to_string(stops_total) +
-               " stops " /* +
+      LOG_INFO(GraphTile::FileSuffix(tile) + " should have " + std::to_string(stations_total) +
+               " stations " /* +
           std::to_string(routes_total) +  " routes and " + std::to_string(pairs_total) +  " stop_pairs"*/);
     }
   }
@@ -280,44 +280,140 @@ std::priority_queue<weighted_tile_t> which_tiles(const ptree& pt, const std::str
       set(value);                                                                                    \
   }
 
-void get_stops(Transit_Fetch& tile,
-               std::unordered_map<std::string, uint64_t>& stops,
-               const GraphId& tile_id,
-               const ptree& response,
-               const AABB2<PointLL>& filter,
-               bool tile_within_one_tz,
-               const std::unordered_map<uint32_t, multi_polygon_type>& tz_polys) {
-  for (const auto& stop_pt : response.get_child("stops")) {
-    const auto& ll_pt = stop_pt.second.get_child("geometry_centroid.coordinates");
+void get_stop_stations(Transit& tile,
+                       std::unordered_map<std::string, uint64_t>& nodes,
+                       std::unordered_map<std::string, uint64_t>& platforms,
+                       const GraphId& tile_id,
+                       const ptree& response,
+                       const AABB2<PointLL>& filter,
+                       bool tile_within_one_tz,
+                       const std::unordered_map<uint32_t, multi_polygon_type>& tz_polys) {
+
+  for (const auto& station_pt : response.get_child("stop_stations")) {
+
+    const auto& ll_pt = station_pt.second.get_child("geometry_centroid.coordinates");
     auto lon = ll_pt.front().second.get_value<float>();
     auto lat = ll_pt.back().second.get_value<float>();
-    if (!filter.Contains({lon, lat})) {
+    std::string onestop_id = station_pt.second.get<std::string>("onestop_id");
+
+    if (!filter.Contains({lon, lat}) || station_pt.second.get_child("stop_egresses").size() == 0 ||
+        station_pt.second.get_child("stop_platforms").size() == 0 ||
+        onestop_id.find('>') != std::string::npos || onestop_id.find('<') != std::string::npos) {
       continue;
     }
-    auto* stop = tile.add_stops();
-    stop->set_lon(lon);
-    stop->set_lat(lat);
-    set_no_null(std::string, stop_pt.second, "onestop_id", "null", stop->set_onestop_id);
-    set_no_null(std::string, stop_pt.second, "name", "null", stop->set_name);
-    stop->set_wheelchair_boarding(stop_pt.second.get<bool>("wheelchair_boarding", true));
-    set_no_null(uint64_t, stop_pt.second, "tags.osm_way_id", 0, stop->set_osm_way_id);
-    GraphId stop_id(tile_id.tileid(), tile_id.level(), stops.size());
-    stop->set_graphid(stop_id);
 
-    auto tz = stop_pt.second.get<std::string>("timezone", "null");
-    if (tz != "null") {
-      stop->set_timezone(tz);
+    // add egresses first.
+    GraphId prev_type_graphid;
+    for (const auto& egress_pt : station_pt.second.get_child("stop_egresses")) {
+      const auto& ll_pt = egress_pt.second.get_child("geometry_centroid.coordinates");
+      auto lon = ll_pt.front().second.get_value<float>();
+      auto lat = ll_pt.back().second.get_value<float>();
+      auto* node = tile.add_nodes();
+
+      node->set_lon(lon);
+      node->set_lat(lat);
+      set_no_null(std::string, egress_pt.second, "onestop_id", "null", node->set_onestop_id);
+      node->set_type(static_cast<uint32_t>(NodeType::kTransitEgress));
+      set_no_null(std::string, egress_pt.second, "name", "null", node->set_name);
+      node->set_wheelchair_boarding(egress_pt.second.get<bool>("wheelchair_boarding", true));
+      set_no_null(uint64_t, egress_pt.second, "osm_way_id", 0, node->set_osm_way_id);
+      node->set_generated(egress_pt.second.get<bool>("generated", true));
+
+      auto traversability = egress_pt.second.get<std::string>("directionality", "null");
+      node->set_traversability(static_cast<uint32_t>(Traversability::kBoth));
+
+      // null (both; can enter and exit the egress)
+      // enter (can enter the egress from street)
+      // exit (can exit egress from street)
+      if (traversability == "enter") {
+        node->set_traversability(static_cast<uint32_t>(Traversability::kForward));
+      } else if (traversability == "exit") {
+        node->set_traversability(static_cast<uint32_t>(Traversability::kBackward));
+      }
+
+      GraphId egress_id(tile_id.tileid(), tile_id.level(), nodes.size());
+      node->set_graphid(egress_id);
+
+      // we want to set the previous id to the first egress in the
+      // list so that when we write to the valhalla tile we know
+      // where to start.
+      if (!prev_type_graphid.Is_Valid()) {
+        prev_type_graphid = egress_id;
+      }
+
+      auto tz = egress_pt.second.get<std::string>("timezone", "null");
+      if (tz != "null") {
+        node->set_timezone(tz);
+      }
+
+      nodes.emplace(node->onestop_id(), egress_id);
+      if (nodes.size() == kMaxGraphId) {
+        LOG_ERROR("Hit the maximum number of nodes allowed and skipping the rest");
+        return;
+      }
     }
 
-    stops.emplace(stop->onestop_id(), stop_id);
-    if (stops.size() == kMaxGraphId) {
-      LOG_ERROR("Hit the maximum number of stops allowed and skipping the rest");
-      break;
+    auto* node = tile.add_nodes();
+    // stations next.
+    node->set_lon(lon);
+    node->set_lat(lat);
+    set_no_null(std::string, station_pt.second, "onestop_id", "null", node->set_onestop_id);
+    node->set_type(static_cast<uint32_t>(NodeType::kTransitStation));
+    set_no_null(std::string, station_pt.second, "name", "null", node->set_name);
+    node->set_wheelchair_boarding(station_pt.second.get<bool>("wheelchair_boarding", true));
+    GraphId station_id(tile_id.tileid(), tile_id.level(), nodes.size());
+    node->set_graphid(station_id);
+
+    auto tz = station_pt.second.get<std::string>("timezone", "null");
+    if (tz != "null") {
+      node->set_timezone(tz);
+    }
+
+    node->set_prev_type_graphid(prev_type_graphid);
+    nodes.emplace(node->onestop_id(), station_id);
+    if (nodes.size() == kMaxGraphId) {
+      LOG_ERROR("Hit the maximum number of nodes allowed and skipping the rest");
+      return;
+    }
+
+    // finally add the platforms
+    for (const auto& platforms_pt : station_pt.second.get_child("stop_platforms")) {
+      const auto& ll_pt = platforms_pt.second.get_child("geometry_centroid.coordinates");
+      auto lon = ll_pt.front().second.get_value<float>();
+      auto lat = ll_pt.back().second.get_value<float>();
+      auto* node = tile.add_nodes();
+
+      node->set_lon(lon);
+      node->set_lat(lat);
+      set_no_null(std::string, platforms_pt.second, "onestop_id", "null", node->set_onestop_id);
+      node->set_type(static_cast<uint32_t>(NodeType::kMultiUseTransitPlatform));
+      set_no_null(std::string, platforms_pt.second, "name", "null", node->set_name);
+      node->set_wheelchair_boarding(platforms_pt.second.get<bool>("wheelchair_boarding", true));
+      GraphId platform_id(tile_id.tileid(), tile_id.level(), nodes.size());
+      node->set_graphid(platform_id);
+
+      auto tz = platforms_pt.second.get<std::string>("timezone", "null");
+      if (tz != "null") {
+        node->set_timezone(tz);
+      }
+
+      node->set_prev_type_graphid(station_id);
+      std::string onestop = node->onestop_id();
+      if (onestop.back() == '<') {
+        onestop.pop_back();
+      }
+
+      nodes.emplace(node->onestop_id(), platform_id);
+      platforms.emplace(onestop, platform_id);
+      if (nodes.size() == kMaxGraphId) {
+        LOG_ERROR("Hit the maximum number of nodes allowed and skipping the rest");
+        return;
+      }
     }
   }
 }
 
-void get_routes(Transit_Fetch& tile,
+void get_routes(Transit& tile,
                 std::unordered_map<std::string, size_t>& routes,
                 const std::unordered_map<std::string, std::string>& websites,
                 const std::unordered_map<std::string, std::string>& short_names,
@@ -326,28 +422,28 @@ void get_routes(Transit_Fetch& tile,
     auto* route = tile.add_routes();
     set_no_null(std::string, route_pt.second, "onestop_id", "null", route->set_onestop_id);
     std::string vehicle_type = route_pt.second.get<std::string>("vehicle_type", "null");
-    Transit_Fetch_VehicleType type = Transit_Fetch_VehicleType::Transit_Fetch_VehicleType_kRail;
+    Transit_VehicleType type = Transit_VehicleType::Transit_VehicleType_kRail;
     if (vehicle_type == "tram" || vehicle_type == "tram_service") {
-      type = Transit_Fetch_VehicleType::Transit_Fetch_VehicleType_kTram;
+      type = Transit_VehicleType::Transit_VehicleType_kTram;
     } else if (vehicle_type == "metro") {
-      type = Transit_Fetch_VehicleType::Transit_Fetch_VehicleType_kMetro;
+      type = Transit_VehicleType::Transit_VehicleType_kMetro;
     } else if (vehicle_type == "rail" || vehicle_type == "suburban_railway" ||
                vehicle_type == "railway_service") {
-      type = Transit_Fetch_VehicleType::Transit_Fetch_VehicleType_kRail;
+      type = Transit_VehicleType::Transit_VehicleType_kRail;
     } else if (vehicle_type == "bus" || vehicle_type == "trolleybus_service" ||
                vehicle_type == "express_bus_service" || vehicle_type == "local_bus_service" ||
                vehicle_type == "bus_service" || vehicle_type == "shuttle_bus" ||
                vehicle_type == "demand_and_response_bus_service" ||
                vehicle_type == "regional_bus_service" || vehicle_type == "coach_service") {
-      type = Transit_Fetch_VehicleType::Transit_Fetch_VehicleType_kBus;
+      type = Transit_VehicleType::Transit_VehicleType_kBus;
     } else if (vehicle_type == "ferry") {
-      type = Transit_Fetch_VehicleType::Transit_Fetch_VehicleType_kFerry;
+      type = Transit_VehicleType::Transit_VehicleType_kFerry;
     } else if (vehicle_type == "cablecar") {
-      type = Transit_Fetch_VehicleType::Transit_Fetch_VehicleType_kCableCar;
+      type = Transit_VehicleType::Transit_VehicleType_kCableCar;
     } else if (vehicle_type == "gondola") {
-      type = Transit_Fetch_VehicleType::Transit_Fetch_VehicleType_kGondola;
+      type = Transit_VehicleType::Transit_VehicleType_kGondola;
     } else if (vehicle_type == "funicular") {
-      type = Transit_Fetch_VehicleType::Transit_Fetch_VehicleType_kFunicular;
+      type = Transit_VehicleType::Transit_VehicleType_kFunicular;
     } else {
       LOG_ERROR("Skipping unsupported vehicle_type: " + vehicle_type + " for route " +
                 route->onestop_id());
@@ -388,7 +484,7 @@ void get_routes(Transit_Fetch& tile,
   }
 }
 
-void get_stop_patterns(Transit_Fetch& tile,
+void get_stop_patterns(Transit& tile,
                        std::unordered_map<std::string, size_t>& shapes,
                        const ptree& response) {
   for (const auto& shape_pt : response.get_child("route_stop_patterns")) {
@@ -421,7 +517,7 @@ struct unique_transit_t {
   std::unordered_map<std::string, size_t> lines;
 };
 
-bool get_stop_pairs(Transit_Fetch& tile,
+bool get_stop_pairs(Transit& tile,
                     unique_transit_t& uniques,
                     const std::unordered_map<std::string, size_t>& shapes,
                     const ptree& response,
@@ -432,8 +528,15 @@ bool get_stop_pairs(Transit_Fetch& tile,
     auto* pair = tile.add_stop_pairs();
 
     // origin
-    pair->set_origin_onestop_id(pair_pt.second.get<std::string>("origin_onestop_id"));
-    auto origin = stops.find(pair->origin_onestop_id());
+    auto origin_id = pair_pt.second.get<std::string>("origin_onestop_id");
+
+    if (origin_id.find('<') == std::string::npos) {
+      pair->set_origin_onestop_id(origin_id + "<");
+    } else {
+      pair->set_origin_onestop_id(origin_id);
+    }
+
+    auto origin = stops.find(origin_id);
     if (origin != stops.cend()) {
       pair->set_origin_graphid(origin->second);
     } else {
@@ -441,8 +544,15 @@ bool get_stop_pairs(Transit_Fetch& tile,
     }
 
     // destination
-    pair->set_destination_onestop_id(pair_pt.second.get<std::string>("destination_onestop_id"));
-    auto destination = stops.find(pair->destination_onestop_id());
+    auto destination_id = pair_pt.second.get<std::string>("destination_onestop_id");
+
+    if (destination_id.find('<') == std::string::npos) {
+      pair->set_destination_onestop_id(destination_id + "<");
+    } else {
+      pair->set_destination_onestop_id(destination_id);
+    }
+
+    auto destination = stops.find(destination_id);
     if (destination != stops.cend()) {
       pair->set_destination_graphid(destination->second);
     } else {
@@ -604,35 +714,6 @@ bool get_stop_pairs(Transit_Fetch& tile,
   return dangles;
 }
 
-void write_pbf(const Transit_Fetch& tile, const boost::filesystem::path& transit_tile) {
-  // check for empty stop pairs and routes.
-  if (tile.stop_pairs_size() == 0 && tile.routes_size() == 0 && tile.shapes_size() == 0) {
-    LOG_WARN(transit_tile.string() + " had no data and will not be stored");
-    return;
-  }
-
-  // write pbf to file
-  if (!boost::filesystem::exists(transit_tile.parent_path())) {
-    boost::filesystem::create_directories(transit_tile.parent_path());
-  }
-  auto size = tile.ByteSize();
-  valhalla::midgard::mem_map<char> buffer;
-  buffer.create(transit_tile.string(), size);
-  if (!tile.SerializeToArray(buffer.get(), size)) {
-    LOG_ERROR("Couldn't write: " + transit_tile.string() + " it would have been " +
-              std::to_string(size));
-  }
-
-  if (tile.routes_size() && tile.stops_size() && tile.stop_pairs_size() && tile.shapes_size()) {
-    LOG_INFO(transit_tile.string() + " had " + std::to_string(tile.stops_size()) + " stops " +
-             std::to_string(tile.routes_size()) + " routes " + std::to_string(tile.shapes_size()) +
-             " shapes " + std::to_string(tile.stop_pairs_size()) + " stop pairs");
-  } else {
-    LOG_INFO(transit_tile.string() + " had " + std::to_string(tile.stop_pairs_size()) +
-             " stop pairs");
-  }
-}
-
 void fetch_tiles(const ptree& pt,
                  std::priority_queue<weighted_tile_t>& queue,
                  unique_transit_t& uniques,
@@ -681,11 +762,11 @@ void fetch_tiles(const ptree& pt,
                             ? "&import_level=" + pt.get<std::string>("import_level")
                             : "";
 
-    Transit_Fetch tile;
+    Transit tile;
     auto file_name = GraphTile::FileSuffix(current);
     file_name = file_name.substr(0, file_name.size() - 3) + "pbf";
-    boost::filesystem::path transit_tile =
-        pt.get<std::string>("mjolnir.transit_dir") + '/' + file_name;
+    boost::filesystem::path transit_tile = pt.get<std::string>("mjolnir.transit_dir") +
+                                           filesystem::path::preferred_separator + file_name;
 
     // tiles are wrote out with .pbf or .pbf.n ext
     uint32_t ext = 0;
@@ -701,26 +782,29 @@ void fetch_tiles(const ptree& pt,
       }
     }
 
-    // pull out all the STOPS (you see what we did there?)
-    std::unordered_map<std::string, uint64_t> stops;
+    // all the nodes...stations, platforms, and egresses
+    std::unordered_map<std::string, uint64_t> nodes;
+    // just the platforms
+    std::unordered_map<std::string, uint64_t> platforms;
     boost::optional<std::string> request =
-        url((boost::format("/api/v1/stops?total=false&per_page=%1%&bbox=%2%,%3%,%4%,%5%") %
-             pt.get<std::string>("per_page") % bbox.minx() % bbox.miny() % bbox.maxx() % bbox.maxy())
+        url((boost::format("/api/v1/stop_stations?total=false&per_page=1000&bbox=%1%,%2%,%3%,%4%") %
+             bbox.minx() % bbox.miny() % bbox.maxx() % bbox.maxy())
                 .str(),
             pt);
     request = *request + import_level;
 
     do {
       // grab some stuff
-      response = curler(*request, "stops");
+      response = curler(*request, "stop_stations");
       // copy stops in, keeping map of stopid to graphid
-      get_stops(tile, stops, current, response, filter, tile_within_one_tz, tz_polys);
+      get_stop_stations(tile, nodes, platforms, current, response, filter, tile_within_one_tz,
+                        tz_polys);
       // please sir may i have some more?
       request = response.get_optional<std::string>("meta.next");
 
     } while (request && (request = *request + api_key));
     // um yeah.. we need these
-    if (stops.size() == 0) {
+    if (nodes.size() == 0) {
       LOG_WARN(transit_tile.string() + " had no stops and will not be stored");
       continue;
     }
@@ -798,11 +882,12 @@ void fetch_tiles(const ptree& pt,
 
     // pull out all SCHEDULE_STOP_PAIRS
     bool dangles = false;
-    for (const auto& stop : stops) {
+    for (const auto& platform : platforms) {
+
       request = url((boost::format("/api/v1/"
                                    "schedule_stop_pairs?active=true&total=false&per_page=%1%&"
                                    "origin_onestop_id=%2%&service_from_date=%3%-%4%-%5%") %
-                     pt.get<std::string>("per_page") % url_encode(stop.first) % utc->tm_year %
+                     pt.get<std::string>("per_page") % url_encode(platform.first) % utc->tm_year %
                      utc->tm_mon % utc->tm_mday)
                         .str(),
                     pt);
@@ -811,7 +896,7 @@ void fetch_tiles(const ptree& pt,
         // grab some stuff
         response = curler(*request, "schedule_stop_pairs");
         // copy pairs in, noting if any dont have stops
-        dangles = get_stop_pairs(tile, uniques, shapes, response, stops, routes) || dangles;
+        dangles = get_stop_pairs(tile, uniques, shapes, response, platforms, routes) || dangles;
         // if stop pairs is large save to a path with an incremented extension
         if (tile.stop_pairs_size() >= 500000) {
           LOG_INFO("Writing " + transit_tile.string());
@@ -873,27 +958,6 @@ std::list<GraphId> fetch(const ptree& pt,
   return dangling;
 }
 
-Transit_Fetch read_pbf(const std::string& file_name, std::mutex& lock) {
-  lock.lock();
-  std::fstream file(file_name, std::ios::in | std::ios::binary);
-  if (!file) {
-    throw std::runtime_error("Couldn't load " + file_name);
-    lock.unlock();
-  }
-  std::string buffer((std::istreambuf_iterator<char>(file)), std::istreambuf_iterator<char>());
-  lock.unlock();
-  google::protobuf::io::ArrayInputStream as(static_cast<const void*>(buffer.c_str()), buffer.size());
-  google::protobuf::io::CodedInputStream cs(
-      static_cast<google::protobuf::io::ZeroCopyInputStream*>(&as));
-  auto limit = std::max(static_cast<size_t>(1), buffer.size() * 2);
-  cs.SetTotalBytesLimit(limit, limit);
-  Transit_Fetch transit;
-  if (!transit.ParseFromCodedStream(&cs)) {
-    throw std::runtime_error("Couldn't load " + file_name);
-  }
-  return transit;
-}
-
 struct dist_sort_t {
   PointLL center;
   Tiles<PointLL> grid;
@@ -918,7 +982,8 @@ void stitch_tiles(const ptree& pt,
   auto tile_name = [&pt](const GraphId& id) {
     auto file_name = GraphTile::FileSuffix(id);
     file_name = file_name.substr(0, file_name.size() - 3) + "pbf";
-    return pt.get<std::string>("mjolnir.transit_dir") + '/' + file_name;
+    return pt.get<std::string>("mjolnir.transit_dir") + filesystem::path::preferred_separator +
+           file_name;
   };
 
   // for each tile
@@ -962,10 +1027,10 @@ void stitch_tiles(const ptree& pt,
         if (neighbor_id != current) {
           auto neighbor_file_name = tile_name(neighbor_id);
           auto neighbor = read_pbf(neighbor_file_name, lock);
-          for (const auto& stop : neighbor.stops()) {
-            auto stop_itr = needed.find(stop.onestop_id());
-            if (stop_itr != needed.cend()) {
-              stop_itr->second.value = stop.graphid();
+          for (const auto& node : neighbor.nodes()) {
+            auto platform_itr = needed.find(node.onestop_id());
+            if (platform_itr != needed.cend()) {
+              platform_itr->second.value = node.graphid();
               ++found;
             }
           }
@@ -1038,13 +1103,14 @@ void stitch(const ptree& pt,
 
 int main(int argc, char** argv) {
   if (argc < 2) {
-    std::cerr << "Usage: " << std::string(argv[0])
-              << " valhalla_config transit_land_url per_page [target_directory] [bounding_box] "
-                 "[transit_land_api_key]"
-              << std::endl;
+    std::cerr
+        << "Usage: " << std::string(argv[0])
+        << " valhalla_config transit_land_url per_page [target_directory] [bounding_box]"
+           "[transit_land_api_key] [import_level] [feed_onestop_id] [import_level] [feed_onestop_id]"
+        << std::endl;
     std::cerr << "Sample: " << std::string(argv[0])
               << " conf/valhalla.json http://transit.land/ 1000 ./transit_tiles "
-                 "-31.56,36.63,-6.18,42.16 transitland-YOUR_KEY_SUFFIX"
+                 "-122.469,37.502,-121.78,38.018 transitland-YOUR_KEY_SUFFIX 4 f-9q9-bart"
               << std::endl;
     return 1;
   }
@@ -1090,7 +1156,7 @@ int main(int argc, char** argv) {
 
   // figure out which transit tiles even exist
   boost::filesystem::recursive_directory_iterator transit_file_itr(
-      pt.get<std::string>("mjolnir.transit_dir") + '/' +
+      pt.get<std::string>("mjolnir.transit_dir") + filesystem::path::preferred_separator +
       std::to_string(TileHierarchy::levels().rbegin()->first));
   boost::filesystem::recursive_directory_iterator end_file_itr;
   std::unordered_set<GraphId> all_tiles;
